--- src/NavBar.css
@@ -1,13 +1,0 @@
-.linkBar {
-  background-color: green;
-  color: black;
-}
-
-.pageList {
-  color: black;
-}
-
-.listName {
-  color: black;
-}
-
